requirements.txt
fastapi
uvicorn[standard]
sqlalchemy==1.4.46
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-decouple